split the endpoint module
extract the client, server and transport parts into their own modules.
