+++ src/protocol/mod.rs
@@ -1,6 +1,11 @@
+pub mod client;
 pub mod codec;
 pub mod endpoint;
 pub mod errors;
 pub mod message;
+pub mod server;
+pub mod transport;
 
-pub use self::endpoint::{Ack, Client, Endpoint, Response, Service, ServiceBuilder};
+pub use self::client::{Ack, Client, Response};
+pub use self::endpoint::Endpoint;
+pub use self::server::{IntoStaticFuture, Service, ServiceBuilder};
